Implemented sign in page, connected to Redux
This should be a good launching point for any application that needs user authentication. I'll leave a tag here, so that when building future applications I can use this as a starter project.

Note: create a new firebase project and copy its config to firebase.js when starting new projects. I already exposed the API key - it's alright for a test project but for future reference I'd rather not leave everything public.